move utils to common package
