+++ src/pururu/application/events/entities.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from enum import Enum
 
-import pururu.utils as utils
+from pururu.common import utils
 from pururu.domain.entities import BotEvent, Attendance, Poll
