+++ src/pururu/common/utils.py
