+++ src/pururu/domain/current_session.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 import pururu.config as config
-import pururu.utils as utils
+from pururu.common import utils
 from pururu.domain.entities import Poll
