+++ src/pururu/infrastructure/adapters/discord/discord_service_adapter.py
@@ -2,7 +2,7 @@ from datetime import timedelta
 
 import discord
 
-import pururu.utils as utils
+from pururu.common import utils
 from pururu.domain.entities import Message, Poll
 from pururu.domain.exceptions import DiscordServiceException
 from pururu.domain.services.discord_service import DiscordInterface
