+++ src/tests/test_infrastructure/test_adapters/test_google_sheets/test_google_sheets_adapter.py
@@ -14,7 +14,7 @@ from unittest.mock import patch, Mock, MagicMock
 
 @patch('pururu.infrastructure.adapters.google_sheets.google_sheets_adapter.gspread')
 @patch('google.oauth2.service_account.Credentials.from_service_account_file')
-@patch('pururu.utils.get_logger')
+@patch('pururu.common.utils.get_logger')
 def set_up(gspread_mock, credentials_mock, logger_mock) -> GoogleSheetsAdapter:
     gs_client_mock = Mock()
     gspread_mock.authorize.return_value = gs_client_mock
